refactor mappers

+++ src/main/java/com/pustovalov/service/PersistenceMatchService.java
@@ -1,11 +1,12 @@
 package com.pustovalov.service;
 
 import com.pustovalov.dao.HibernateMatchDao;
-import com.pustovalov.dto.response.MatchResultDto;
 import com.pustovalov.dto.response.StoredMatchesDto;
 import com.pustovalov.entity.Match;
+import com.pustovalov.exception.MatchAlreadyPersistException;
 import com.pustovalov.service.mapper.MatchMapper;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 public class PersistenceMatchService {
@@ -35,12 +36,15 @@ public class PersistenceMatchService {
     return instance;
   }
 
-  public MatchResultDto save(UUID matchId) {
-    Match match = ongoingMatchService.getMatch(matchId);
-    hibernateMatchDao.save(match);
-    ongoingMatchService.delete(matchId);
-
-    return mapper.toMatchResultDto(match);
+  public void save(UUID matchId) {
+    Optional<Match> match = ongoingMatchService.getMatch(matchId);
+    if (match.isEmpty() && hibernateMatchDao.isExist(matchId)) {
+      throw new MatchAlreadyPersistException(
+          String.format("The match(id = %s) has already been saved", matchId));
+    } else {
+      hibernateMatchDao.save(match.get());
+      ongoingMatchService.delete(matchId);
+    }
   }
 
   public StoredMatchesDto findAll(int page) {
@@ -51,7 +55,7 @@ public class PersistenceMatchService {
     int offset = page * LIMIT;
     List<Match> matches = hibernateMatchDao.findAll(offset, LIMIT);
 
-    return mapper.toStoredMatchesDto(matches, totalPages, page);
+    return mapper.toStoredMatchesDto(matches, totalPages, page, page == totalPages);
   }
 
   public StoredMatchesDto findAll(int page, String name) {
@@ -62,14 +66,14 @@ public class PersistenceMatchService {
     int offset = page * LIMIT;
     List<Match> matches = hibernateMatchDao.findByPlayerName(offset, LIMIT, name);
 
-    return mapper.toStoredMatchesDto(matches, name, totalPages, page);
+    return mapper.toStoredMatchesDto(matches, name, totalPages, page, page == totalPages);
   }
 
   private int getTotalPages() {
-    return (int) (hibernateMatchDao.getRowsAmount() / LIMIT);
+    return (int) Math.ceil((double) hibernateMatchDao.getRowsAmount() / LIMIT);
   }
 
   private int getTotalPages(String name) {
-    return (int) (hibernateMatchDao.getRowsAmount(name) / LIMIT);
+    return (int) Math.ceil((double) hibernateMatchDao.getRowsAmount(name) / LIMIT);
   }
 }
